+++ docs/misc/changelog.rst
@@ -107,6 +107,7 @@ New Features:
 - Added support for n-step returns for off-policy algorithms via the `n_steps` parameter
 - Added ``NStepReplayBuffer`` that allows to compute n-step returns without additional memory requirement (and without for loops)
 - Added Gymnasium v1.2 support
+- Added Prioritized Approximation Loss
 
 Bug Fixes:
 ^^^^^^^^^^